requirements.txt
streamlit
sqlalchemy
bcrypt
pypdf
google-genai
python-dotenv
pandas
openpyxl
psycopg2-binary
